requirements.txt
pandas 
sqlalchemy 
psycopg2-binary
flask
$#fKcdXPg4@ue8AW
